A Lot of Changes
-Change Necrozma form species names
-Room Service + Eject Pack
-Fix Extreme Evoboost
-Fix Clangorous Soulblaze not Sound
-Update move tutors
-Higher pitch mega cry always
-Doubles AI Classes
-Add Camomons
-Fix Saveblock1 struct offsets
-Number only naming screen
-Shiny frontier spreads
-Visual type calc for move menu
-Unbound Battle Frontier DNS
-Fix DNS windows updating every frame
-Begin more battle facilities
-Multi frontier specials
-Fix Random32() being signed
-Fix ABILITY macro usage
-Acupressure AI targeting
-Multi battle mugshot fix
-Egg Move Relearner

+++ include/new/ram_locs_battle.h
@@ -175,7 +175,9 @@ typedef u8 Battle_Buffer_T[0x200];
 #define ShakerData (((u8*) 0x203E02C))
 #define ForceSwitchHelper (*((u8*) 0x203E02E))
 #define AbilityPopUpHelper (*((u8*) 0x203E02F))
-#define BackupHWord (*((u8*) 0x203E030))
+#define BackupHWord (*((u16*) 0x203E030))
+
+//Don't go past 0x203E048
 
 #define FIRST_NEW_BATTLE_RAM_LOC ((u8*) 0x203E020)
 #define LAST_NEW_BATTLE_RAM_LOC ((u8*) 0x203E032)
